backend/requirements.txt
asgiref==3.3.1
dj-database-url==0.5.0
Django==3.1.5
django-better-admin-arrayfield==1.4.2
django-cors-headers==3.6.0
django-jazzmin==2.4.3
django-rest-framework==0.1.0
django-webpack-loader==0.7.0
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
Pillow==8.1.0
psycopg2-binary==2.8.6
PyJWT==1.7.1
python-http-client==3.3.1
pytz==2020.5
sendgrid==6.4.8
sqlparse==0.4.1
starkbank-ecdsa==1.1.0
